pyproject.toml
[tool.poetry]
name = "courses-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
black = "^24.8.0"
isort = "^5.13.2"
ipython = "^8.26.0"
django-filter = "^24.3"
coverage = "^7.6.1"
drf-yasg = "^1.21.7"
setuptools = "^73.0.1"
django-cors-headers = "^4.4.0"
stripe = "^10.8.0"
requests = "^2.32.3"
rest-framework-simplejwt = "^0.0.2"
celery = "^5.4.0"
eventlet = "^0.36.1"
redis = "^5.0.8"
django-celery-beat = "^2.7.0"
django = "^5.1.1"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
